% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_pkg}
\alias{build_pkg}
\title{build various versions of a package}
\usage{
build_pkg(.pkg = ".", types = c("source", "binary"),
  repository = NULL, origin = NULL, supplement_version = FALSE,
  overwrite = TRUE)
}
\arguments{
\item{.pkg}{path to package}

\item{repository}{repository name being built for}

\item{origin}{package source}

\item{supplement_version}{whether to add additional version info (unix timestamp)}
}
\description{
build various versions of a package
}
\details{
supplementing version can be done whenever a build occurs that does not
correspond to a formal release/tag. This will automatically add information
about the git hash (if available), as well as incrememnt the version number
with a unix timestamp that corresponds to the last git hash (if present) or
the current system time, if git is not present.
}
