% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release.R
\name{create_tagged_repo}
\alias{create_tagged_repo}
\title{create a cranlike repo that contains the tagged package}
\usage{
create_tagged_repo(.dir, repo_name = "pkgpub", ...)
}
\arguments{
\item{.dir}{directory to create the repo in}

\item{repo_name}{repository name to specify in the description}

\item{...}{additional arguments passed to build_pkg}
}
\description{
create a cranlike repo that contains the tagged package
}
