% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.R
\name{insert_packages}
\alias{insert_packages}
\title{insert packages into a drat repo}
\usage{
insert_packages(pkgs, repository, archive = TRUE, ...)
}
\arguments{
\item{pkgs}{list of packages to insert}

\item{repository}{repository folder to insert}

\item{archive}{whether to archive the packages after insertion to prevent duplicates}

\item{...}{parameters passed to `drat::archivePackages`}
}
\description{
insert packages into a drat repo
}
\details{
repository folder should most likely correspond to the name
of the repository set in the pkg metadata, however
realistically, the time that they need to match when
repos is set in a user session in the `getOption('repos')` field
}
