% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-repo.R
\name{new_cranlike_repo}
\alias{new_cranlike_repo}
\title{new cranlike repo will create a cranlike repo}
\usage{
new_cranlike_repo(dir = getwd())
}
\arguments{
\item{dir}{directory to create}
}
\description{
new cranlike repo will create a cranlike repo
}
