% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exit.R
\name{skip_remaining_if_dev_tag}
\alias{skip_remaining_if_dev_tag}
\title{skip remaining drone CI steps if tag is dev tag}
\usage{
skip_remaining_if_dev_tag()
}
\description{
skip remaining drone CI steps if tag is dev tag
}
