% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_version.R
\name{tag_version}
\alias{tag_version}
\title{add a git tag based on the pkg version}
\usage{
tag_version(.dir = getwd(), .dirty = FALSE, .fetch = TRUE, .push = TRUE)
}
\arguments{
\item{.dir}{dir of repo}

\item{.dirty}{whether to tag given the repo is in a dirty state}

\item{.fetch}{run a git fetch first to get tags}

\item{.push}{push created tag}
}
\description{
add a git tag based on the pkg version
}
