% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_pkg}
\alias{build_pkg}
\title{build various versions of a package}
\usage{
build_pkg(.pkgdir = ".", types = c("source", "binary"),
  repository = NULL, origin = NULL, addl_meta = NULL,
  supplement_version = FALSE, overwrite = TRUE, ...)
}
\arguments{
\item{.pkgdir}{path to package}

\item{types}{types of package to build}

\item{repository}{repository name being built for}

\item{origin}{package source}

\item{addl_meta}{additional metadata}

\item{supplement_version}{whether to add additional version info (unix timestamp)}

\item{overwrite}{overwrite fields already present when adding fields}

\item{...}{parameters to pass to pkgbuild}
}
\description{
build various versions of a package
}
\details{
supplementing version can be done whenever a build occurs that does not
correspond to a formal release/tag. This will automatically add information
about the git hash (if available), as well as incrememnt the version number
with a unix timestamp that corresponds to the last git hash (if present) or
the current system time, if git is not present.
}
