% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release.R
\name{create_tagged_repo}
\alias{create_tagged_repo}
\title{create a cranlike repo that contains the tagged package}
\usage{
create_tagged_repo(.dir = "/tmp")
}
\arguments{
\item{.dir}{directory to create the repo in}
}
\description{
create a cranlike repo that contains the tagged package
}
